# pilot: fix n^2 scaling behavior with gateways