--- a/pilot/pkg/model/push_context.go
+++ b/pilot/pkg/model/push_context.go
@@ -91,13 +91,13 @@ type PushContext struct {
 	QuotaSpecBinding []Config `json:"-"`
 
 	// VirtualService related
-	// this contains all the virtual services with exportTo "." and current namespace
-	privateVirtualServicesByNamespace map[string][]Config
 	// This contains all virtual services visible to this namespace extracted from
-	// exportTos that explicitly contained this namespace.
-	virtualServicesExportedToNamespace map[string][]Config
-	// This contains all virtual services whose exportTo is "*"
-	publicVirtualServices []Config
+	// exportTos that explicitly contained this namespace. The keys are namespace,gateway.
+	virtualServicesExportedToNamespaceByGateway map[string]map[string][]Config
+	// this contains all the virtual services with exportTo "." and current namespace. The keys are namespace,gateway.
+	privateVirtualServicesByNamespaceAndGateway map[string]map[string][]Config
+	// This contains all virtual services whose exportTo is "*", keyed by gateway
+	publicVirtualServicesByGateway map[string][]Config
 
 	// destination rules are of three types:
 	//  namespaceLocalDestRules: all public/private dest rules pertaining to a service defined in a given namespace
@@ -443,21 +443,21 @@ func init() {
 func NewPushContext() *PushContext {
 	// TODO: detect push in progress, don't update status if set
 	return &PushContext{
-		publicServices:                     []*Service{},
-		privateServicesByNamespace:         map[string][]*Service{},
-		servicesExportedToNamespace:        map[string][]*Service{},
-		publicVirtualServices:              []Config{},
-		privateVirtualServicesByNamespace:  map[string][]Config{},
-		virtualServicesExportedToNamespace: map[string][]Config{},
-		namespaceLocalDestRules:            map[string]*processedDestRules{},
-		exportedDestRulesByNamespace:       map[string]*processedDestRules{},
-		sidecarsByNamespace:                map[string][]*SidecarScope{},
-		envoyFiltersByNamespace:            map[string][]*EnvoyFilterWrapper{},
-		gatewaysByNamespace:                map[string][]Config{},
-		allGateways:                        []Config{},
-		ServiceByHostnameAndNamespace:      map[host.Name]map[string]*Service{},
-		ProxyStatus:                        map[string]map[string]ProxyPushStatus{},
-		ServiceAccounts:                    map[host.Name]map[int][]string{},
+		publicServices:                              []*Service{},
+		privateServicesByNamespace:                  map[string][]*Service{},
+		servicesExportedToNamespace:                 map[string][]*Service{},
+		publicVirtualServicesByGateway:              map[string][]Config{},
+		privateVirtualServicesByNamespaceAndGateway: map[string]map[string][]Config{},
+		virtualServicesExportedToNamespaceByGateway: map[string]map[string][]Config{},
+		namespaceLocalDestRules:                     map[string]*processedDestRules{},
+		exportedDestRulesByNamespace:                map[string]*processedDestRules{},
+		sidecarsByNamespace:                         map[string][]*SidecarScope{},
+		envoyFiltersByNamespace:                     map[string][]*EnvoyFilterWrapper{},
+		gatewaysByNamespace:                         map[string][]Config{},
+		allGateways:                                 []Config{},
+		ServiceByHostnameAndNamespace:               map[host.Name]map[string]*Service{},
+		ProxyStatus:                                 map[string]map[string]ProxyPushStatus{},
+		ServiceAccounts:                             map[host.Name]map[int][]string{},
 	}
 }
 
@@ -529,8 +529,6 @@ func virtualServiceDestinations(v *networking.VirtualService) []*networking.Dest
 // GatewayServices returns the set of services which are referred from the proxy gateways.
 func (ps *PushContext) GatewayServices(proxy *Proxy) []*Service {
 	svcs := ps.Services(proxy)
-	// gateway set.
-	gateways := map[string]bool{}
 	// host set.
 	hostsFromGateways := map[string]struct{}{}
 
@@ -541,19 +539,16 @@ func (ps *PushContext) GatewayServices(proxy *Proxy) []*Service {
 	}
 
 	for _, gw := range proxy.MergedGateway.GatewayNameForServer {
-		gateways[gw] = true
-	}
-	log.Debugf("GatewayServices: gateway %v has following gw resources:%v", proxy.ID, gateways)
+		for _, vsConfig := range ps.VirtualServicesForGateway(proxy, gw) {
+			vs, ok := vsConfig.Spec.(*networking.VirtualService)
+			if !ok { // should never happen
+				log.Errorf("Failed in getting a virtual service: %v", vsConfig.Labels)
+				return svcs
+			}
 
-	for _, vsConfig := range ps.VirtualServices(proxy, gateways) {
-		vs, ok := vsConfig.Spec.(*networking.VirtualService)
-		if !ok { // should never happen
-			log.Errorf("Failed in getting a virtual service: %v", vsConfig.Labels)
-			return svcs
-		}
-
-		for _, d := range virtualServiceDestinations(vs) {
-			hostsFromGateways[d.Host] = struct{}{}
+			for _, d := range virtualServiceDestinations(vs) {
+				hostsFromGateways[d.Host] = struct{}{}
+			}
 		}
 	}
 
@@ -603,46 +598,11 @@ func (ps *PushContext) Services(proxy *Proxy) []*Service {
 // VirtualServices lists all virtual services bound to the specified gateways
 // This replaces store.VirtualServices. Used only by the gateways
 // Sidecars use the egressListener.VirtualServices().
-func (ps *PushContext) VirtualServices(proxy *Proxy, gateways map[string]bool) []Config {
-	configs := make([]Config, 0)
-	out := make([]Config, 0)
-
-	// filter out virtual services not reachable
-	// First add private virtual services and explicitly exportedTo virtual services
-	if proxy == nil {
-		for _, virtualSvcs := range ps.privateVirtualServicesByNamespace {
-			configs = append(configs, virtualSvcs...)
-		}
-	} else {
-		configs = append(configs, ps.privateVirtualServicesByNamespace[proxy.ConfigNamespace]...)
-		configs = append(configs, ps.virtualServicesExportedToNamespace[proxy.ConfigNamespace]...)
-	}
-	// Second add public virtual services.
-	configs = append(configs, ps.publicVirtualServices...)
-
-	for _, cfg := range configs {
-		rule := cfg.Spec.(*networking.VirtualService)
-		if len(rule.Gateways) == 0 {
-			// This rule applies only to IstioMeshGateway
-			if _, ok := gateways[constants.IstioMeshGateway]; ok {
-				out = append(out, cfg)
-			}
-		} else {
-			for _, g := range rule.Gateways {
-				// note: Gateway names do _not_ use wildcard matching, so we do not use Name.Matches here
-				if _, ok := gateways[resolveGatewayName(g, cfg.ConfigMeta)]; ok {
-					out = append(out, cfg)
-					break
-				} else if _, ok := gateways[g]; ok && g == constants.IstioMeshGateway {
-					// "mesh" gateway cannot be expanded into FQDN
-					out = append(out, cfg)
-					break
-				}
-			}
-		}
-	}
-
-	return out
+func (ps *PushContext) VirtualServicesForGateway(proxy *Proxy, gateway string) []Config {
+	res := ps.privateVirtualServicesByNamespaceAndGateway[proxy.ConfigNamespace][gateway][:]
+	res = append(res, ps.virtualServicesExportedToNamespaceByGateway[proxy.ConfigNamespace][gateway]...)
+	res = append(res, ps.publicVirtualServicesByGateway[gateway]...)
+	return res
 }
 
 // getSidecarScope returns a SidecarScope object associated with the
@@ -947,9 +907,9 @@ func (ps *PushContext) updateContext(
 			return err
 		}
 	} else {
-		ps.privateVirtualServicesByNamespace = oldPushContext.privateVirtualServicesByNamespace
-		ps.virtualServicesExportedToNamespace = oldPushContext.virtualServicesExportedToNamespace
-		ps.publicVirtualServices = oldPushContext.publicVirtualServices
+		ps.virtualServicesExportedToNamespaceByGateway = oldPushContext.virtualServicesExportedToNamespaceByGateway
+		ps.privateVirtualServicesByNamespaceAndGateway = oldPushContext.privateVirtualServicesByNamespaceAndGateway
+		ps.publicVirtualServicesByGateway = oldPushContext.publicVirtualServicesByGateway
 	}
 
 	if destinationRulesChanged {
@@ -1106,9 +1066,10 @@ func (ps *PushContext) initAuthnPolicies(env *Environment) error {
 
 // Caches list of virtual services
 func (ps *PushContext) initVirtualServices(env *Environment) error {
-	ps.privateVirtualServicesByNamespace = map[string][]Config{}
-	ps.virtualServicesExportedToNamespace = map[string][]Config{}
-	ps.publicVirtualServices = []Config{}
+	ps.virtualServicesExportedToNamespaceByGateway = map[string]map[string][]Config{}
+	ps.privateVirtualServicesByNamespaceAndGateway = map[string]map[string][]Config{}
+	ps.publicVirtualServicesByGateway = map[string][]Config{}
+
 	virtualServices, err := env.List(gvk.VirtualService, NamespaceAll)
 	if err != nil {
 		return err
@@ -1134,79 +1095,28 @@ func (ps *PushContext) initVirtualServices(env *Environment) error {
 
 	// convert all shortnames in virtual services into FQDNs
 	for _, r := range vservices {
-		rule := r.Spec.(*networking.VirtualService)
-		// resolve top level hosts
-		for i, h := range rule.Hosts {
-			rule.Hosts[i] = string(ResolveShortnameToFQDN(h, r.ConfigMeta))
-		}
-		// resolve gateways to bind to
-		for i, g := range rule.Gateways {
-			if g != constants.IstioMeshGateway {
-				rule.Gateways[i] = resolveGatewayName(g, r.ConfigMeta)
-			}
-		}
-		// resolve host in http route.destination, route.mirror
-		for _, d := range rule.Http {
-			for _, m := range d.Match {
-				for i, g := range m.Gateways {
-					if g != constants.IstioMeshGateway {
-						m.Gateways[i] = resolveGatewayName(g, r.ConfigMeta)
-					}
-				}
-			}
-			for _, w := range d.Route {
-				if w.Destination != nil {
-					w.Destination.Host = string(ResolveShortnameToFQDN(w.Destination.Host, r.ConfigMeta))
-				}
-			}
-			if d.Mirror != nil {
-				d.Mirror.Host = string(ResolveShortnameToFQDN(d.Mirror.Host, r.ConfigMeta))
-			}
-		}
-		// resolve host in tcp route.destination
-		for _, d := range rule.Tcp {
-			for _, m := range d.Match {
-				for i, g := range m.Gateways {
-					if g != constants.IstioMeshGateway {
-						m.Gateways[i] = resolveGatewayName(g, r.ConfigMeta)
-					}
-				}
-			}
-			for _, w := range d.Route {
-				if w.Destination != nil {
-					w.Destination.Host = string(ResolveShortnameToFQDN(w.Destination.Host, r.ConfigMeta))
-				}
-			}
-		}
-		//resolve host in tls route.destination
-		for _, tls := range rule.Tls {
-			for _, m := range tls.Match {
-				for i, g := range m.Gateways {
-					if g != constants.IstioMeshGateway {
-						m.Gateways[i] = resolveGatewayName(g, r.ConfigMeta)
-					}
-				}
-			}
-			for _, w := range tls.Route {
-				if w.Destination != nil {
-					w.Destination.Host = string(ResolveShortnameToFQDN(w.Destination.Host, r.ConfigMeta))
-				}
-			}
-		}
+		resolveVirtualServiceShortnames(r.Spec.(*networking.VirtualService), r.ConfigMeta)
 	}
 
 	for _, virtualService := range vservices {
 		ns := virtualService.Namespace
 		rule := virtualService.Spec.(*networking.VirtualService)
+		gwNames := getGatewayNames(rule, virtualService.ConfigMeta)
 		if len(rule.ExportTo) == 0 {
 			// No exportTo in virtualService. Use the global default
 			// We only honor ., *
 			if ps.defaultVirtualServiceExportTo[visibility.Private] {
+				if _, f := ps.privateVirtualServicesByNamespaceAndGateway[ns]; !f {
+					ps.privateVirtualServicesByNamespaceAndGateway[ns] = map[string][]Config{}
+				}
 				// add to local namespace only
-				ps.privateVirtualServicesByNamespace[ns] =
-					append(ps.privateVirtualServicesByNamespace[ns], virtualService)
+				for _, gw := range gwNames {
+					ps.privateVirtualServicesByNamespaceAndGateway[ns][gw] = append(ps.privateVirtualServicesByNamespaceAndGateway[ns][gw], virtualService)
+				}
 			} else if ps.defaultVirtualServiceExportTo[visibility.Public] {
-				ps.publicVirtualServices = append(ps.publicVirtualServices, virtualService)
+				for _, gw := range gwNames {
+					ps.publicVirtualServicesByGateway[gw] = append(ps.publicVirtualServicesByGateway[gw], virtualService)
+				}
 			}
 		} else {
 			exportToMap := make(map[visibility.Instance]bool)
@@ -1217,7 +1127,9 @@ func (ps *PushContext) initVirtualServices(env *Environment) error {
 			// if vs has exportTo *, make public and ignore all other exportTos
 			// if vs has exportTo ., replace with current namespace
 			if exportToMap[visibility.Public] {
-				ps.publicVirtualServices = append(ps.publicVirtualServices, virtualService)
+				for _, gw := range gwNames {
+					ps.publicVirtualServicesByGateway[gw] = append(ps.publicVirtualServicesByGateway[gw], virtualService)
+				}
 				continue
 			} else if exportToMap[visibility.None] {
 				// not possible
@@ -1226,13 +1138,22 @@ func (ps *PushContext) initVirtualServices(env *Environment) error {
 				// . or other namespaces
 				for exportTo := range exportToMap {
 					if exportTo == visibility.Private || string(exportTo) == ns {
-						// exportTo with same namespace is effectively private
-						ps.privateVirtualServicesByNamespace[ns] =
-							append(ps.privateVirtualServicesByNamespace[ns], virtualService)
+						if _, f := ps.privateVirtualServicesByNamespaceAndGateway[ns]; !f {
+							ps.privateVirtualServicesByNamespaceAndGateway[ns] = map[string][]Config{}
+						}
+						// add to local namespace only
+						for _, gw := range gwNames {
+							ps.privateVirtualServicesByNamespaceAndGateway[ns][gw] = append(ps.privateVirtualServicesByNamespaceAndGateway[ns][gw], virtualService)
+						}
 					} else {
-						// exportTo is a specific target namespace
-						ps.virtualServicesExportedToNamespace[string(exportTo)] =
-							append(ps.virtualServicesExportedToNamespace[string(exportTo)], virtualService)
+						if _, f := ps.virtualServicesExportedToNamespaceByGateway[string(exportTo)]; !f {
+							ps.virtualServicesExportedToNamespaceByGateway[string(exportTo)] = map[string][]Config{}
+						}
+						// add to local namespace only
+						for _, gw := range gwNames {
+							ps.virtualServicesExportedToNamespaceByGateway[string(exportTo)][gw] =
+								append(ps.virtualServicesExportedToNamespaceByGateway[string(exportTo)][gw], virtualService)
+						}
 					}
 				}
 			}
@@ -1240,6 +1161,24 @@ func (ps *PushContext) initVirtualServices(env *Environment) error {
 	}
 
 	return nil
+}
+
+var meshGateways = []string{constants.IstioMeshGateway}
+
+func getGatewayNames(vs *networking.VirtualService, meta ConfigMeta) []string {
+	if len(vs.Gateways) == 0 {
+		return meshGateways
+	}
+	res := make([]string, 0, len(vs.Gateways))
+	for _, g := range vs.Gateways {
+		if g == constants.IstioMeshGateway {
+			res = append(res, constants.IstioMeshGateway)
+		} else {
+			name := resolveGatewayName(g, meta)
+			res = append(res, name)
+		}
+	}
+	return res
 }
 
 func (ps *PushContext) initDefaultExportMaps() {