Fixes https://github.com/istio/istio/issues/25116

Against current master
```
name                               old time/op        new time/op        delta
InitPushContext/gateways-6               4.27ms ± 1%        4.03ms ± 1%   -5.59%  (p=0.008 n=5+5)
InitPushContext/gateways-shared-6         662µs ± 1%         465µs ± 1%  -29.75%  (p=0.008 n=5+5)
InitPushContext/empty-6                   302ns ± 2%         298ns ± 2%     ~     (p=0.222 n=5+5)
InitPushContext/telemetry-6               298ns ± 2%         298ns ± 3%     ~     (p=0.635 n=5+5)
InitPushContext/virtualservice-6          168ns ± 4%         171ns ± 2%     ~     (p=0.183 n=5+5)
RouteGeneration/gateways-6                387ms ± 1%         112ms ± 1%  -71.03%  (p=0.008 n=5+5)
RouteGeneration/gateways-shared-6         462ms ± 3%         453ms ± 3%     ~     (p=0.222 n=5+5)
RouteGeneration/empty-6                   841µs ± 2%         835µs ± 5%     ~     (p=0.310 n=5+5)
RouteGeneration/telemetry-6               851µs ± 3%         851µs ± 3%     ~     (p=1.000 n=5+5)
RouteGeneration/virtualservice-6         2.01ms ± 2%        1.99ms ± 2%     ~     (p=0.310 n=5+5)
```

Currently, we store all of the virtual services in a list. Then for each route (1 per gateway) we traverse over each virtual service: https://github.com/istio/istio/blob/f467d22f165ed39d996a45049ffa2d26234596f3/pilot/pkg/networking/core/v1alpha3/gateway.go#L218. Basically we have O(VS*GW) scaling. This ultimately leads to enormous config gen times for RDS, GC thrashing, etc - a single generation for RDS for *one* ingress with 2k gateways/virtual services took over 30s seconds

This change optimizes the model to make it O(VS+GW). We store virtual services keyed by gateway, so we can do a quick lookup. This ultimately helps with non-gateway as well, as we can lookup the mesh gateway quickly as well.

Interestingly, the code is currently logically setup like this, just not performance wise - VirtualService takes in a set of gateways to get virtual services for, but the set always has exactly one element.